backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
python-dotenv
python-multipart